# apps/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kedai Bunda API",
        "description": "API untuk Sistem Kasir dan Manajemen Stok Kedai Bunda",
        "contact": {
            "email": "admin@kedaibunda.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Menampilkan daftar user (hanya super_admin)",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Daftar user berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data user"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Super Admin"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "admin@kedaibunda.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "super_admin"
                                                    },
                                                    "aktif": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Menampilkan detail user",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Berhasil mengambil detail user"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Super Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@kedaibunda.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "super_admin"
                                                },
                                                "aktif": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "User tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mengupdate user (hanya super_admin)",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data user yang ingin diupdate",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Admin Updated"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin.updated@kedaibunda.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "super_admin",
                                            "admin"
                                        ],
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "User berhasil diupdate"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Menghapus user (hanya super_admin)",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "User berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak dapat menghapus akun sendiri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Tidak dapat menghapus akun sendiri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/profil": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update profil sendiri",
                "operationId": "3cca232ba5f59621e9539f68523027b1",
                "requestBody": {
                    "description": "Data profil yang ingin diupdate",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Nama Baru"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "email.baru@kedaibunda.com"
                                    },
                                    "password_lama": {
                                        "description": "Required jika ingin ganti password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldpassword"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Profil berhasil diupdate"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password lama tidak sesuai",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Password lama tidak sesuai"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}